AMP-700: move DashboardResult from web to model
