+++ src/main/java/edu/indiana/dlib/amppd/controller/DashboardController.java
@@ -8,8 +8,8 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import edu.indiana.dlib.amppd.model.DashboardResult;
 import edu.indiana.dlib.amppd.service.DashboardService;
-import edu.indiana.dlib.amppd.web.DashboardResult;
 
 @CrossOrigin(origins = "*", allowedHeaders = "*")
 @RestController
